// webapp/src/api/model/secondaryAttribute.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { BinaryExpressionTree } from './binaryExpressionTree';

export interface SecondaryAttribute {
  calculationFormula: BinaryExpressionTree;
  consumable: boolean;
  id?: string;
  name: string;
  shortName: string;
}